A major rewrite of GraphSim.
After including require.js the code was separated in many modules. Another effect is that definition of variables with a global namespace are now restircted to the graphSim module.

The code is still not as clean and legible as it should be. But lesson learned, never start with a first little commit what seems to be a little rewrite ;-).

Fixes #3